TowerUpgrade, TowerSell

+++ Assets/Script/Projectile.cs
@@ -6,9 +6,9 @@ public class Projectile : MonoBehaviour
 {
     private Movement2D movement2D;
     private Transform target;
-    private int damage;
+    private float damage;
 
-    public void SetUp(Transform target, int damage)
+    public void SetUp(Transform target, float damage)
     {
         movement2D = GetComponent<Movement2D>();
         this.target = target;
